+++ src/main.py
@@ -2,12 +2,19 @@ import requests
 import sys
 import json
 import os
-import readRepoCommits
-import markdown
 
-#System arguments
-#python {File.py} {access token} {blog name} {repository name}
-#arguments = sys.argv
+import readRepoCommits
+import getReadmeContents
+
+# System arguments
+# For local test purpose
+# python {File.py} {access token} {blog name} {repository name}
+# arguments = sys.argv
+# blogAccessToken = arguments[1]
+# blogName = arguments[2]
+# repoName = arguments[3]
+# repoOwner = arguments[4]
+# gitAccessToken = arguments[5]
 
 blogAccessToken = os.environ["INPUT_ACCESSTOKEN"]
 blogName = os.environ["INPUT_BLOGNAME"]
@@ -16,45 +23,14 @@ repoOwner = str(Repo_Name).split('/')[0]
 repoName = str(Repo_Name).split('/')[1]
 gitAccessToken = os.environ["INPUT_GITHUBTOKEN"]
 
-# arguments = sys.argv
-# blogAccessToken = arguments[1]
-# blogName = arguments[2]
-# repoName = arguments[3]
-# repoOwner = arguments[4]
-# gitAccessToken = arguments[5]
 
 global contents
 contents = ''
 
-def convert_md_to_html(md_content):
-    html_content = markdown.markdown(md_content)
-    return html_content
-
-def get_readme_from_github(owner, repo):
-    url = f'https://api.github.com/repos/{owner}/{repo}/contents/README.md'
-    
-    try:
-        response = requests.get(url)
-        response.raise_for_status()
-        readme_data = response.json()
-        
-        # Check if the response contains a 'content' key
-        if 'content' in readme_data:
-            import base64
-            readme_content = base64.b64decode(readme_data['content']).decode('utf-8')
-            return readme_content
-        else:
-            print("README content not found in the response.")
-            return None
-    except requests.exceptions.RequestException as e:
-        print(f"Error fetching README from GitHub: {e}")
-        return None
-
 def contents_generator():
-    #ReadMe 읽고 맨 위에 쓰고
     global contents
     
-    contents += convert_md_to_html(get_readme_from_github(repoOwner , repoName))
+    contents += getReadmeContents.convert_md_to_html(getReadmeContents.get_readme_from_github(repoOwner , repoName))
     
     commits = readRepoCommits.get_commits(repoOwner, repoName, gitAccessToken)
     commitCounter = 1
@@ -146,5 +122,5 @@ def check_postExist():
             post_blog()
             break
             
-#post_blog()
-check_postExist()
+if __name__ == "__main__":      
+    check_postExist()
